// MS_rosterfrontend/projects/client/src/lib/view-models/add-client-details.ts
export interface AddClientDetails {
    clientId?: number;
    salutation?: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    dateOfBirth?: string;
    maritalStatus?: string;
    mobileNo?: string;
    gender?: number;
    emailId?: string;
    address?: string;
    locationId?: number;
    employeeId?: number;
    fullName?: string;
    status?:boolean;
}